# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Plateforme de déploiement autonome pour Docker Swarm.",
        "title": "Hivemind API",
        "contact": {
            "name": "Équipe technique Hivemind",
            "email": "issadicko78@gmail.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/services/{id}/exec": {
            "get": {
                "description": "Upgrades to a WebSocket and attaches an interactive TTY exec session to one of the service's containers. Client→server text messages are tagged by a leading byte: '0'=stdin data, '1'=JSON resize {cols,rows}. Server→client messages are raw terminal output. Authenticate with ?token=\u003caccess_token\u003e (WebSocket cannot carry an Authorization header). Requires the Admin role.",
                "tags": [
                    "supervision"
                ],
                "summary": "Interactive container shell (WebSocket, Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target container ID (from /services/{id}/tasks)",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Shell command (default /bin/sh)",
                        "name": "cmd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_orange_hivemind_internal_adapters_api_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_orange_hivemind_internal_adapters_api_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_orange_hivemind_internal_adapters_api_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "service not deployed",
                        "schema": {
                            "$ref": "#/definitions/github_com_orange_hivemind_internal_adapters_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_orange_hivemind_internal_adapters_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}